third_party/blink/web_tests/fast/forms/calendar-picker/date-picker-year-view-resize.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../resources/common.js"></script>
<script src="../resources/picker-common.js"></script>
<script src="../calendar-picker/resources/calendar-picker-common.js"></script>
</head>
<body>

<input type=date id=date value="2019-09-10">

<script>
let t = async_test('Test year list view size');

function test1() {
  var dateViewHeight = internals.pagePopupWindow.global.picker.height();
  clickMonthPopupButton();
  assert_equals(dateViewHeight, internals.pagePopupWindow.global.picker.height());
}

t.step(() => {
  openPicker(document.getElementById('date'))
  .then(t.step_func_done(test1));
});
</script>
</body>
</html>
